requirements.txt
wheel>=0.22
pyyaml>=3.10
docopt>=0.6.1